+++ src/entities/CatEstatusInstalacion.ts
@@ -5,7 +5,7 @@ import { applySchema } from 'src/common/apply-schema.decorator';
 @Index('IX_CatEstatusInstalacion_Estatus', ['estatus'])
 @Entity('CatEstatusInstalacion')
 export class CatEstatusInstalacion {
-  @PrimaryGeneratedColumn({ type: 'int', name: 'Id' })
+  @PrimaryGeneratedColumn({ type: 'bigint', name: 'Id' })
   id: number;
 
   @Column('varchar', { name: 'Nombre', length: 50 })
